Allow AccessibleByteArrayOutputStream to shrink
AccessibleByteArrayOutputStream currently can only grow
or stay the same size. This could be a problem if it is used for a rare
large item, but then used only for small items - the buffer would stay
large forever, consuming memory. Since an application may hold many of
these buffers via their use particularly in Cipher, the memory impact
could be significant.

This simple shrinkage algorithm shrinks the buffer if it has been 'too
large' N times in a row. N starts at 1, and grows exponentially up to a
limit of 1024 as re-growths are triggered. N starts low to enable
reacting quickly, but as more growths are observed it increases to avoid
churn. It is capped to avoid degrading into 'never-shrink' behaviour
over the long term.

Other ideas were considered, like keeping track of the last N used sizes
and making decisions based on those values, but this algorithm has a
tiny footprint (2 ints, could be shorts) and low computation overhead.