Reduce build time by 35%
The gradle tasks coverage_exec and test_exec correspond to the
underlying CMake tasks coverage and check respectively. In CMake,
the coverage task depends on check. Because of how the gradle task
dependencies are wired up we run these tasks separately and end up
running the CMake check task twice. The check target is responsible for
running all of our unit tests which is by far the longest portion of our
build which is why not doing that twice cuts our build time by 35%. This
avoids the redundancy by re-arranging the dependency graph in gradle to
run coverage once and generate both unit test and coverage reports out
of it.