Match SunRsaSign's behavior when setting null params (#356)
We're stuck in a bit of a compatibility pickle -- when null parameters
are specified, Sun's behavior is to throw an exception but BC's behavior
is to carry on silently. So, which one should we choose? We've decided
to go with Sun's behavior because even though this change will mean
_more_ cases where an exception is thrown than before (which could
amount to a breaking change in some cases),
InvalidAlgorithmParameterException [is a checked exception][1], so the
compiler will force callers to handle it in some form. If
InvalidAlgorithmParameterException were an unchecked RuntimeException,
this change would not be advisable, as callers may not be properly
handling that case.

[1]: https://docs.oracle.com/javase/8/docs/api/java/security/InvalidAlgorithmParameterException.html