Use `LibCryptoRng` instead of `DEFAULT` in Random benchmark
Following reversal of PR #376 in PR #388, using the ACCP SecureRandom
algorithm `DEFAULT` (which is an alias of `LibCryptoRng`) yields lower
performance in multi-threaded settings.
See #376 for details.

This PR is meant to use the algorithm `LibCryptoRng` in the benchmark,
instead of its alias `DEFAULT`. This solves the performance drop in
multi-threaded settings.

This also corresponds to the most common use of ACCP SecureRandom.
Indeed, if ACCP is the first security provider and if ACCP SecureRandom
 is registered, then when instantiating SecureRandom as
 `new SecureRandom()`, the algorithm `LibCryptoRng` will be selected.