Apply the rule of 3/5/0 to JByteArrayCritical et al.
Apply the [rule of 3/5/0](https://en.cppreference.com/w/cpp/language/rule_of_three)
to JByteArrayCritical, SimpleBuffer, JBinaryBlob, JIOBlobs.
This ensures that these objects cannot be copied.

Before, if they were copied (e.g., due to being passed as argument to a
function), the destructor would be called twice, which would release the
critical region twice or which may release the critical region while
another part of the code is still using it.