Remove CPP coverage target from default coverage target
Mainly, we want to remove this to unblock our CI. This task is currently
breaking all Mac CI dimensions (and possibly others). The root cause
seems related to some mismatches in the related software dependencies
used to create the coverage report.

More broadly, this coverage report is not in a state of good health and
should not bere-introduced to the CI until we take the time addressing
these issues: Redundant unit test execution due to dependency hierarchy,
ignored errors in the coverage report resulting in poor coverage of the
coverage data, surfacing the coverage reporting in a useful manner.