Fix classname field of AES KeyGenerator Service
In PR #280 the package hierarchy for this class was flattened, but this
line of code was not updated to reflect this. In practice, this all
worked and continued to work since we don't 'useReflection' to
instantiate this crypto service. However, we do send up the classname
into the JCA so whatever that uses it for may have been broken until
now.