Add VERBOSE build flag (#451)
*Description of changes:*

The build logs are getting excessive to the point that Github and
CodeBuild are struggling to store/render them. This adds a new
VERBOSE build flag to require opt-in for detailed logging. For starters,
we use this to toggle `CMAKE_VERBOSE_MAKEFILE` for the JNI and AWS-LC
builds which was previously ON by default.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.