Update test to handle a bug in Signature.verify() implementation by SunEC EdDSASignature until JDK 21 (#494)
*Issue #332 *

*Description of changes:*

According to the `Signature.verify()`
[interface](https://github.com/openjdk/jdk/blob/jdk-17%2B35/src/java.base/share/classes/java/security/Signature.java#L782-L786),
if anything is wrong with the signature the implementation should throw
a `SignatureException`, which should ideally apply for the case of
`null` signature as well.

But until JDK 20, the `sun.security.ec.ed.EdDSASignature`
[implementation](https://github.com/openjdk/jdk/blob/jdk-20%2B35/src/jdk.crypto.ec/share/classes/sun/security/ec/ed/EdDSASignature.java#L213-L215)
skipped signature verification and returned a `false` instead, when its
`message` object was null (default value until assigned via an update()
call).
Bug report: https://bugs.openjdk.org/browse/JDK-8300399

In JDK 21, the
[implementation](https://github.com/openjdk/jdk/blob/jdk-21%2B35/src/jdk.crypto.ec/share/classes/sun/security/ec/ed/EdDSASignature.java#L213-L215)
was
[fixed](https://github.com/openjdk/jdk/commit/b317658d69a477df04ded3cc2e107970f8a6e20d)
to perform verification for empty (null) message, and thereby comply
with Signature.verify() interface to throw `SignatureException` for a
`null` signature.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.