Aes wrap not padded (#498)
*Issue #493*

*Description of changes:*

This PR adds a new CipherSpi implementation, AesKeyWrapSpi. This
cipher can be instantiated as AES with mode KW, and NoPadding. AWS-LC
only provides one-shot functions for key wrapping and unwrapping, so we
buffer any incremental data fed to the Cipher until finalization, at
which point we pass all buffered data over the JNI and call AWS-LC. AES
KW is defined in [RFC-3394][1], which we rely on for output buffer
sizing and other implementation-adjacent details/input validation.

As described in [this document][2], NIST provides Known Answer Tests for
AES KW. We pull down [these test vectors][3] and include `gzip`'s of
them in the repo. We then add AesKwKatTest to read these test cases in,
and execute them against ACCP. This commit follows prior art for KAT
tests in AesKwpKatTest.

[1]: https://datatracker.ietf.org/doc/html/rfc3394
[2]:
https://csrc.nist.gov/CSRC/media/Projects/Cryptographic-Algorithm-Validation-Program/documents/mac/KWVS.pdf
[3]:
https://csrc.nist.gov/CSRC/media/Projects/Cryptographic-Algorithm-Validation-Program/documents/mac/kwtestvectors.zip

---------

Co-authored-by: Sergii Gubenko <s_gubenko@apple.com>
Co-authored-by: Alex Weibel <alexw91@gmail.com>