Add POSIX permissions checks to Loader (#505)
*Issue #, if available:*
CryptoAlg-3511

*Description of changes:*
Assertions around the POSIX permissions and owner of the temporary
directory created by `Loader` have been added to prevent a race
condition between creation of the directory and use of it for loading
the ACCP native library.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.